Implement vertical scrollbar for Window, including mouse wheel handling. Organize the OpenGL context creating routine. (This will not work on some old GPU.) Make Rect support zero width and height.

+++ src/ImGui/Port/IO/Mouse.cs
@@ -135,7 +135,7 @@
             get { return mousePos != lastMousePos; }
         }
 
-        public float MouseWheel { get; set; }
+        public short MouseWheel { get; set; }
 
         #endregion
